+++ fota_sender/fota_sender.c
@@ -29,8 +29,6 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "fota_crc_tool.h"
-
 static const mira_net_config_t net_config = {
     .pan_id = 0x12345678,
     .key = { 0xaa,
@@ -130,7 +128,8 @@ PROCESS_THREAD(fota_sender_proc, ev, data)
     static mira_size_t i;
     static mira_size_t j;
     static uint8_t buffer[32];
-    static uint32_t crc_state;
+    static mira_crc_ctx_t crc_ctx;
+    uint32_t crc_value;
 
     PROCESS_BEGIN();
 
@@ -168,7 +167,7 @@ PROCESS_THREAD(fota_sender_proc, ev, data)
         PROCESS_WAIT_WHILE(mira_fota_is_working());
 
         /* Write some arbitrary data, which could be a firmware */
-        fota_crc_init(&crc_state);
+        mira_crc_init(&crc_ctx);
 
         image_size = 10000 + (mira_random_generate() % 1000);
         for (i = 0; i < image_size; i += 32) {
@@ -179,7 +178,7 @@ PROCESS_THREAD(fota_sender_proc, ev, data)
             for (j = 0; j < block_size; j++) {
                 buffer[j] = (i + j + version_no) % 77;
             }
-            fota_crc_update(&crc_state, buffer, block_size);
+            mira_crc_update(&crc_ctx, buffer, block_size);
             if (mira_fota_write(i, buffer, block_size) != MIRA_SUCCESS) {
                 printf("ERROR: mira_fota_write failed\n");
                 break;
@@ -187,8 +186,8 @@ PROCESS_THREAD(fota_sender_proc, ev, data)
             PROCESS_WAIT_WHILE(mira_fota_is_working());
         }
 
-        if (mira_fota_write_header(image_size, fota_crc_get(&crc_state), 0, 0, version_no) !=
-            MIRA_SUCCESS) {
+        mira_crc_get(&crc_ctx, &crc_value);
+        if (mira_fota_write_header(image_size, crc_value, 0, 0, version_no) != MIRA_SUCCESS) {
             printf("ERROR: mira_fota_write_header failed\n");
             break;
         }